products-server/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
cloudinary==1.44.1
flask==3.1.1
flask-cors==6.0.1
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.14.1
urllib3==2.5.0
werkzeug==3.1.3
